requirements.txt
asgiref==3.5.0
certifi==2020.6.20
Django==3.2.11
data-spec-validator==1.5.0
djangorestframework==3.13.1
django-cors-headers==3.7.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-redis==4.11.0
django-simpleui==2022.1
django-simple-history==3.0.0
django-storages[google]==1.12.3
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
google==3.0.0
google-cloud-logging==3.0.0
google-cloud-storage==1.30.0
mysqlclient==1.4.6
Pillow==7.1.2
pytz==2021.3
redis==3.5.0
sqlparse==0.4.2
typing-extensions==4.0.1
